Fixed previous pull request comments and did a spell check (#5)
* commit on maste

* commit

* classId -> periodId

* added row mappers to everything else

* nothing

* Checked spelling everywhere+fixed Henry's comment on previous pull request

* deleted comments at end of each file

* Update Event.java

* Update EventRowMapper.java

+++ src/main/java/dev/mentalspace/wafflecone/databaseobject/Preference.java
@@ -1,12 +1,12 @@
 package dev.mentalspace.wafflecone.databaseobject;
 
 public class Preference {
-    public long preferenceId;
-    public long studentId;
-    public int assignmentOrder;
-    public int startType;
-    public int breakLength;
-    public int breakFrequency;
+    public Long preferenceId;
+    public Long studentId;
+    public Integer assignmentOrder;
+    public Integer startType;
+    public Long breakLength;
+    public Long breakFrequency;
 
     public Preference() {
     }
